добавляет хук useMediaPlayer
