+++ src/app/components/media-player/media-player.tsx
@@ -3,7 +3,7 @@ import { IPlayer } from "interfaces";
 import cn from "classnames";
 import { VideoPlayer } from "app/components/video-player";
 import { Modal } from "app/components/modal";
-import { useMediaPlayerStore } from "./use-media-player-store";
+import { useMediaPlayer } from "./use-media-player";
 import { StartButton } from "./media-player.start-button";
 import "./media-player.scss";
 
@@ -12,12 +12,9 @@ type MediaPlayerProps = {
 }
 
 export const MediaPlayer = (props: MediaPlayerProps) => {
-    const { player } = props;
     const [isOpen, setIsOpen] = React.useState(false);
     const [isRolledUp, setIsRolledUp] = React.useState(false);
-    const store = useMediaPlayerStore();
-
-    player.addStore(store);
+    const { isPlaying, player } = useMediaPlayer(props.player);
 
     const handleStartPlay = () => {
         setIsOpen(true);
@@ -39,7 +36,7 @@ export const MediaPlayer = (props: MediaPlayerProps) => {
                 </div>
             )}
             <Modal onClose={handleClose} className={cn(isRolledUp && "media-player__rolled")} isOpen={isOpen}>
-                <VideoPlayer player={player} state={store.state} isRolledUp={isRolledUp} toggleRollup={toggleRollup} />
+                <VideoPlayer player={player} isPlaying={isPlaying} isRolledUp={isRolledUp} toggleRollup={toggleRollup} />
             </Modal>
         </div>
     )
