+++ src/app/components/video-player/video-player.tsx
@@ -5,14 +5,13 @@ import * as Types from "./video-player.types";
 import "./video-player.scss";
 
 export const VideoPlayer: React.FC<Types.VideoPlayerProps> = (props) => {
-    const { player, state, isRolledUp, toggleRollup } = props;
-    const videoRef = React.useRef<HTMLVideoElement>(null);
+    const { player, isPlaying, isRolledUp, toggleRollup } = props;
 
-    const { isPlaying } = state;
-
-    React.useLayoutEffect(() => {
-        player.assignElement(videoRef.current);
-    }, [player]);
+    const initPlayer = (node: HTMLVideoElement | null) => {
+        if(node) {
+            player.assignElement(node);
+        }
+    }
 
     const handlePlay = () => {
         player.play();
@@ -24,7 +23,7 @@ export const VideoPlayer: React.FC<Types.VideoPlayerProps> = (props) => {
 
     return (
         <div>
-            <video poster={player.getCurrentPosterSrc()} src={player.getCurrentVideoSrc()} ref={videoRef} className={cn("video-player__video", isRolledUp && "rolled")} />
+            <video poster={player.getCurrentPosterSrc()} src={player.getCurrentVideoSrc()} ref={initPlayer} className={cn("video-player__video", isRolledUp && "rolled")} />
             <div className="video-player__controls">
                 <Button className={cn("video-player__button", isRolledUp ? "unroll" : "roll" )} onClick={toggleRollup} content="" />
                 <Button className={cn("video-player__button", isPlaying ? "pause" : "play" )} onClick={isPlaying ? handlePause : handlePlay} content="" />
